Move test code to test package
